Remove pandas from analysis
